+++ packages/be-ts-lib/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 🔖 [3.0.15] (2026-02-20)
+
+### 🏗️ Changed
+
+- Updated dependency: now requires `@aneuhold/core-ts-api-lib@^3.0.15`.
+
 ## 🔖 [3.0.14] (2026-02-17)
 
 ### 🏗️ Changed
@@ -180,6 +186,7 @@ No direct code changes; version bump for compatibility with new major versions o
 
 <!-- Link References -->
 
+[3.0.15]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.0.14...be-ts-lib-v3.0.15
 [3.0.14]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.0.13...be-ts-lib-v3.0.14
 [3.0.13]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.0.12...be-ts-lib-v3.0.13
 [3.0.12]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.0.11...be-ts-lib-v3.0.12